+++ battleship/prompts/standard_prompts.py
